[UPDATED] Typo Fixed

+++ 0x1E-search_algorithms/100-jump.c
@@ -13,29 +13,29 @@
 
 int jump_search(int *array, size_t size, int value)
 {
-	double result = 0;
-	int index = 0;
+	int result = 0;
+	unsigned int index = 0;
 	int i = 0;
 
 	if (!array || !value)
 		return (-1);
 
-	result = sqrt((double)size);
+	result = sqrt(size);
 
 	while (array[index] <= value)
 	{
 		printf("Value checked array[%d] = [%d]\n", index, array[index]);
-		index += (int)result;
-		if (array[index] >= value || index > (int)size)
+		index += result;
+		if (array[index] >= value || index > size)
 		{
 			printf("Value found between indexes [%d] and [%d]\n",
-					(index - (int)result), index);
-			for (i = index - (int)result; array[i] <= value; i++)
+					(index - result), index);
+			for (i = index - result; array[i] <= value; i++)
 			{
 				printf("Value checked array[%d] = [%d]\n", i, array[i]);
 				if (array[i] == value)
 					return (i);
-				if (index >= (int)size)
+				if (index >= size)
 					return (-1);
 			}
 		}
